Convert to blocklist with config instead of matching parent name
The first attempt at this rule matched the 'parent' in the MemberExpression if
it ended with "Entity". This a) requires consistent naming of sequelize entity
instances throughout the codebase and b) is specific to our (Source Health's)
codebase.

We felt it would be better to try to match every call to any method called
`create()`, `update()`, etc and allow the developer to either disable false
positives on a case-by case basic, or also to provide a list of common patterns
that can be ignored - e.g. in our codebase we often make calls like
`fooService.create()` or `operations.update()` that are nothing to do with
Sequelize.